project/requirements.txt
streamlit>=1.32.0
pdfplumber>=0.11.0
anthropic>=0.25.0
gspread>=6.0.0
google-auth>=2.29.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
pandas>=2.2.0
